requirements.txt
# API and environment variables
python-dotenv>=1.0.0
requests>=2.31.0

# Image and video processing
Pillow>=10.0.0
opencv-python==4.8.1.78

# LangGraph and dependencies
langgraph>=0.3.21
langchain>=0.1.5
langchain-core>=0.1.16
langchain-openai>=0.0.5
langchain-community>=0.0.10
typing-extensions>=4.5.0
pydantic>=2.0.0
langsmith>=0.0.69
deepseek>=1.0.0
tavily-python>=0.2.2
json-repair>=0.6.0
numpy>=1.24.0
pytest>=7.0.0
pytest-mock>=3.11.0 

langgraph>=0.0.34
langchain>=0.1.0
langchain-openai>=0.0.2

pandas>=1.3.5
fastparquet
pytubefix
youtube-transcript-api

scenedetect
